video/imagetek_i4100.cpp: Fix regression with buffered sprite RAM, Updates - Fix notes position - Reduce duplicated notes
metro/metro.cpp: Split blzntrnd and gstrik2* into separated driver file via different hardware configurations
