+++ src/devices/video/imagetek_i4100.cpp
@@ -376,7 +376,9 @@ void imagetek_i4100_device::device_start()
 	save_item(NAME(m_crtc_horz));
 	save_item(NAME(m_crtc_vert));
 	save_item(NAME(m_sprite_count));
+	save_item(NAME(m_sprite_count_latch));
 	save_item(NAME(m_sprite_priority));
+	save_item(NAME(m_sprite_priority_latch));
 	save_item(NAME(m_sprite_color_code));
 	save_item(NAME(m_sprite_xoffset));
 	save_item(NAME(m_sprite_yoffset));
@@ -421,7 +423,9 @@ void imagetek_i4100_device::device_reset()
 	m_rombank = 0;
 	m_crtc_unlock = false;
 	m_sprite_count = 0;
+	m_sprite_count_latch = 0;
 	m_sprite_priority = 0;
+	m_sprite_priority_latch = 0;
 	m_sprite_xoffset = 0;
 	m_sprite_yoffset = 0;
 	m_sprite_color_code = 0;
@@ -587,8 +591,13 @@ void imagetek_i4100_device::tiletable_w(offs_t offset, uint16_t data, uint16_t m
  * 0.w  ---- ---- ---- ----     Number Of Sprites To Draw
  *
  ************************************************************/
-uint16_t imagetek_i4100_device::sprite_count_r() { return m_sprite_count; }
-void imagetek_i4100_device::sprite_count_w(offs_t offset, uint16_t data, uint16_t mem_mask) { COMBINE_DATA(&m_sprite_count); }
+uint16_t imagetek_i4100_device::sprite_count_r() { return m_sprite_count_latch; }
+void imagetek_i4100_device::sprite_count_w(offs_t offset, uint16_t data, uint16_t mem_mask)
+{
+	COMBINE_DATA(&m_sprite_count_latch);
+	if (!m_spriteram_buffered)
+		m_sprite_count = m_sprite_count_latch;
+}
 
 /*************************************************************
  *
@@ -600,8 +609,13 @@ void imagetek_i4100_device::sprite_count_w(offs_t offset, uint16_t data, uint16_
  *      ---- ---- ---4 3210     Sprites Masked Number
  *
  *************************************************************/
-uint16_t imagetek_i4100_device::sprite_priority_r() { return m_sprite_priority; }
-void imagetek_i4100_device::sprite_priority_w(offs_t offset, uint16_t data, uint16_t mem_mask) { COMBINE_DATA(&m_sprite_priority); }
+uint16_t imagetek_i4100_device::sprite_priority_r() { return m_sprite_priority_latch; }
+void imagetek_i4100_device::sprite_priority_w(offs_t offset, uint16_t data, uint16_t mem_mask)
+{
+	COMBINE_DATA(&m_sprite_priority);
+	if (!m_spriteram_buffered)
+		m_sprite_priority = m_sprite_priority_latch;
+}
 
 /*************************************************************
  *
@@ -829,52 +843,6 @@ void imagetek_i4100_device::blt_write(int const tmap, const offs_t offs, u16 con
 	}
 //  logerror("%s : Blitter %X] %04X <- %04X & %04X\n", machine().describe_context(), tmap, offs, data, mask);
 }
-
-/***************************************************************************
-
-
-                                    Blitter
-
-    [ Registers ]
-
-        Offset:     Value:
-
-        0.l         Destination Tilemap      (1,2,3)
-        4.l         Blitter Data Address     (byte offset into the gfx ROMs)
-        8.l         Destination Address << 7 (byte offset into the tilemap)
-
-        The Blitter reads a byte and looks at the most significative
-        bits for the opcode, while the remaining bits define a value
-        (usually how many bytes to write). The opcode byte may be
-        followed by a number of other bytes:
-
-            76------            Opcode
-            --543210            N
-            (at most N+1 bytes follow)
-
-
-        The blitter is designed to write every other byte (e.g. it
-        writes a byte and skips the next). Hence 2 blits are needed
-        to fill a tilemap (first even, then odd addresses)
-
-    [ Opcodes ]
-
-            0       Copy the following N+1 bytes. If the whole byte
-                    is $00: stop and generate an IRQ
-
-            1       Fill N+1 bytes with a sequence, starting with
-                    the  value in the following byte
-
-            2       Fill N+1 bytes with the value in the following
-                    byte
-
-            3       Skip N+1 bytes. If the whole byte is $C0:
-                    skip to the next row of the tilemap (+0x200 bytes)
-                    but preserve the column passed at the start of the
-                    blit (destination address % 0x200)
-
-
-***************************************************************************/
 
 
 // TODO: clean this up
@@ -1280,6 +1248,26 @@ void imagetek_i4100_device::draw_sprites(screen_device &screen, bitmap_rgb32 &bi
 }
 
 
+/***************************************************************************
+
+                        Tilemaps: Tiles Set & Window
+
+    Each entry in the Tiles Set RAM uses 2 words to specify a starting
+    tile code and a color code. This adds 16 consecutive tiles with
+    that color code to the set of available tiles.
+
+        Offset:     Bits:                   Value:
+
+        0.w         fedc ---- ---- ----
+                    ---- ba98 7654 ----     Color Code*
+                    ---- ---- ---- 3210     Code High Bits
+
+        2.w                                 Code Low Bits
+
+* 00-ff, but on later chips supporting it, xf means 256 color tile and palette x
+
+***************************************************************************/
+
 inline u8 imagetek_i4100_device::get_tile_pix(u16 code, u8 x, u8 y, bool const big, u32 &pix)
 {
 	// Use code as an index into the tiles set table
@@ -1342,26 +1330,6 @@ inline u8 imagetek_i4100_device::get_tile_pix(u16 code, u8 x, u8 y, bool const b
 			return 0;
 	}
 }
-
-/***************************************************************************
-
-                        Tilemaps: Tiles Set & Window
-
-    Each entry in the Tiles Set RAM uses 2 words to specify a starting
-    tile code and a color code. This adds 16 consecutive tiles with
-    that color code to the set of available tiles.
-
-        Offset:     Bits:                   Value:
-
-        0.w         fedc ---- ---- ----
-                    ---- ba98 7654 ----     Color Code*
-                    ---- ---- ---- 3210     Code High Bits
-
-        2.w                                 Code Low Bits
-
-* 00-ff, but on later chips supporting it, xf means 256 color tile and palette x
-
-***************************************************************************/
 
 void imagetek_i4100_device::draw_tilemap(screen_device &screen, bitmap_rgb32 &bitmap, const rectangle &cliprect, u32 flags, u32 const pcode,
 		int sx, int sy, int wx, int wy, bool const big, int const layer)
@@ -1470,6 +1438,10 @@ void imagetek_i4100_device::screen_eof(int state)
 			set_irq(m_vblank_irq_level);
 
 		if (m_spriteram_buffered)
+		{
 			m_spriteram->copy();
+			m_sprite_count = m_sprite_count_latch;
+			m_sprite_priority = m_sprite_priority_latch;
+		}
 	}
 }
