+++ src/devices/video/imagetek_i4100.h
@@ -123,7 +123,9 @@ protected:
 	u16 m_crtc_horz;
 	u16 m_crtc_vert;
 	u16 m_sprite_count;
+	u16 m_sprite_count_latch;
 	u16 m_sprite_priority;
+	u16 m_sprite_priority_latch;
 	u16 m_sprite_xoffset,m_sprite_yoffset;
 	u16 m_sprite_color_code;
 	u8 m_layer_priority[3];
